Refactor Community component: update import path, enhance layout, and improve post handling functionality
