+++ Frontend/src/components/Community.jsx
@@ -1,0 +1,255 @@
+import React, { useState } from "react";
+import Layout from "./Layout";
+
+const Community = () => {
+  const [posts, setPosts] = useState([
+    {
+      id: 1,
+      author: "RahulCitizen",
+      title: "Pothole Issue on MG Road - Update Needed",
+      content: "Has anyone filed a complaint about the pothole near the bus stop? It's been 2 weeks now.",
+      category: "Infrastructure",
+      upvotes: 24,
+      downvotes: 2,
+      comments: 8,
+      time: "2 hours ago",
+      userVote: null
+    },
+    {
+      id: 2,
+      author: "MumbaiResident",
+      title: "Great Response from Municipality!",
+      content: "Filed a complaint about street lights last week. They fixed it in 3 days. Really impressed with the response time.",
+      category: "Success Story",
+      upvotes: 45,
+      downvotes: 0,
+      comments: 12,
+      time: "5 hours ago",
+      userVote: null
+    },
+    {
+      id: 3,
+      author: "GreenWarrior",
+      title: "Garbage Collection Schedule Changes",
+      content: "Notice from municipal corporation about new garbage collection timings in our area. Morning collection shifted to 7 AM.",
+      category: "Announcement",
+      upvotes: 18,
+      downvotes: 3,
+      comments: 5,
+      time: "1 day ago",
+      userVote: null
+    }
+  ]);
+
+  const [newPost, setNewPost] = useState({
+    title: "",
+    content: "",
+    category: "General"
+  });
+
+  const [showNewPost, setShowNewPost] = useState(false);
+
+  const categories = ["General", "Infrastructure", "Water Supply", "Electricity", "Sanitation", "Success Story", "Announcement"];
+
+  const handleVote = (postId, voteType) => {
+    setPosts(posts.map(post => {
+      if (post.id === postId) {
+        let newUpvotes = post.upvotes;
+        let newDownvotes = post.downvotes;
+        let newUserVote = voteType;
+
+        if (post.userVote === voteType) {
+          // Remove vote if clicking same vote
+          newUserVote = null;
+          if (voteType === 'up') newUpvotes--;
+          else newDownvotes--;
+        } else if (post.userVote) {
+          // Change vote
+          if (post.userVote === 'up') {
+            newUpvotes--;
+            newDownvotes++;
+          } else {
+            newDownvotes--;
+            newUpvotes++;
+          }
+        } else {
+          // New vote
+          if (voteType === 'up') newUpvotes++;
+          else newDownvotes++;
+        }
+
+        return { ...post, upvotes: newUpvotes, downvotes: newDownvotes, userVote: newUserVote };
+      }
+      return post;
+    }));
+  };
+
+  const handleCreatePost = (e) => {
+    e.preventDefault();
+    if (newPost.title.trim() && newPost.content.trim()) {
+      const post = {
+        id: posts.length + 1,
+        author: "You",
+        title: newPost.title,
+        content: newPost.content,
+        category: newPost.category,
+        upvotes: 1,
+        downvotes: 0,
+        comments: 0,
+        time: "Just now",
+        userVote: 'up'
+      };
+      setPosts([post, ...posts]);
+      setNewPost({ title: "", content: "", category: "General" });
+      setShowNewPost(false);
+    }
+  };
+
+  const getCategoryColor = (category) => {
+    const colors = {
+      "Infrastructure": "bg-blue-100 text-blue-800",
+      "Water Supply": "bg-cyan-100 text-cyan-800",
+      "Electricity": "bg-yellow-100 text-yellow-800",
+      "Sanitation": "bg-green-100 text-green-800",
+      "Success Story": "bg-emerald-100 text-emerald-800",
+      "Announcement": "bg-purple-100 text-purple-800",
+      "General": "bg-gray-100 text-gray-800"
+    };
+    return colors[category] || "bg-gray-100 text-gray-800";
+  };
+
+  return (
+    <Layout>
+      <div className="space-y-8 w-full max-w-7xl mx-auto">
+        {/* Header */}
+        <div className="flex justify-between items-center">
+          <div>
+            <h1 className="text-4xl font-bold text-gray-900 mb-2">Community Forum</h1>
+            <p className="text-gray-700 text-base">
+              Connect with fellow citizens, share experiences, and help each other with civic issues.
+            </p>
+          </div>
+          <button
+            onClick={() => setShowNewPost(!showNewPost)}
+            className="bg-green-600 hover:bg-green-700 text-white px-4 py-2 rounded-md shadow-sm flex items-center gap-2 font-medium transition-colors"
+          >
+            Create Post
+          </button>
+        </div>
+
+        {/* Create New Post Form */}
+        {showNewPost && (
+          <div className="bg-white rounded-lg shadow-md border border-gray-200">
+            <div className="p-6">
+              <h3 className="text-xl font-semibold text-gray-900 mb-4">Create New Post</h3>
+              <form onSubmit={handleCreatePost}>
+                <div className="mb-4">
+                  <select
+                    value={newPost.category}
+                    onChange={(e) => setNewPost({...newPost, category: e.target.value})}
+                    className="w-full max-w-xs px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-transparent bg-white text-gray-900"
+                  >
+                    {categories.map(cat => (
+                      <option key={cat} value={cat}>{cat}</option>
+                    ))}
+                  </select>
+                </div>
+                <input
+                  type="text"
+                  placeholder="Post title..."
+                  value={newPost.title}
+                  onChange={(e) => setNewPost({...newPost, title: e.target.value})}
+                  className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-transparent mb-4 bg-white text-gray-900"
+                  required
+                />
+                <textarea
+                  placeholder="What's on your mind?"
+                  value={newPost.content}
+                  onChange={(e) => setNewPost({...newPost, content: e.target.value})}
+                  className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-green-500 focus:border-transparent h-32 mb-4 bg-white text-gray-900 resize-none"
+                  required
+                />
+                <div className="flex gap-2">
+                  <button type="submit" className="bg-green-600 hover:bg-green-700 text-white px-4 py-2 rounded-md font-medium transition-colors">
+                    Post
+                  </button>
+                  <button type="button" onClick={() => setShowNewPost(false)} className="px-4 py-2 border border-gray-300 text-gray-900 hover:bg-gray-50 hover:border-gray-400 rounded-md font-medium transition-colors">
+                    Cancel
+                  </button>
+                </div>
+              </form>
+            </div>
+          </div>
+        )}
+
+        {/* Posts List */}
+        <div className="space-y-4">
+          {posts.map(post => (
+            <div key={post.id} className="bg-white rounded-lg shadow-md border border-gray-200 hover:shadow-lg transition-shadow">
+              <div className="p-6">
+                <div className="flex items-start gap-4">
+                  {/* Vote Section */}
+                  <div className="flex flex-col items-center min-w-[60px]">
+                    <button
+                      onClick={() => handleVote(post.id, 'up')}
+                      className={`p-1 rounded hover:bg-gray-100 ${
+                        post.userVote === 'up' ? 'text-green-600' : 'text-gray-500'
+                      }`}
+                    >
+                      ▲
+                    </button>
+                    <span className="text-sm font-semibold text-gray-700">
+                      {post.upvotes - post.downvotes}
+                    </span>
+                    <button
+                      onClick={() => handleVote(post.id, 'down')}
+                      className={`p-1 rounded hover:bg-gray-100 ${
+                        post.userVote === 'down' ? 'text-red-600' : 'text-gray-500'
+                      }`}
+                    >
+                      ▼
+                    </button>
+                  </div>
+
+                  {/* Post Content */}
+                  <div className="flex-1">
+                    <div className="flex items-center gap-2 mb-2">
+                      <span className={`px-2 py-1 rounded-full text-xs font-medium ${getCategoryColor(post.category)}`}>
+                        {post.category}
+                      </span>
+                      <span className="text-sm text-gray-500">
+                        Posted by u/{post.author} • {post.time}
+                      </span>
+                    </div>
+                    
+                    <h3 className="text-lg font-semibold text-gray-900 mb-2">
+                      {post.title}
+                    </h3>
+                    
+                    <p className="text-gray-700 mb-3">
+                      {post.content}
+                    </p>
+
+                    <div className="flex items-center gap-4 text-sm text-gray-500">
+                      <button className="flex items-center gap-1 hover:text-gray-700">
+                        💬 {post.comments} comments
+                      </button>
+                      <button className="flex items-center gap-1 hover:text-gray-700">
+                        📤 Share
+                      </button>
+                      <button className="flex items-center gap-1 hover:text-gray-700">
+                        💾 Save
+                      </button>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          ))}
+        </div>
+      </div>
+    </Layout>
+  );
+};
+
+export default Community;
